HW02: Исправление замечаний к PR
